ACN: REF Vector to use double values.

+++ Core/src/main/java/com/polytech4A/cuttingstock/core/model/Vector.java
@@ -33,12 +33,12 @@ public class Vector {
     /**
      * Horizontal value.
      */
-    private int x;
+    private double x;
 
     /**
      * Vertical value.
      */
-    private int y;
+    private double y;
 
     /**
      * Create a instance of a 2D vector.
@@ -46,28 +46,28 @@ public class Vector {
      * @param x horizontal value.
      * @param y vertical value.
      */
-    public Vector(int x, int y) {
+    public Vector(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public void set(int x, int y) {
+    public void set(double x, double y) {
         this.x = x;
         this.y = y;
     }
